requirements.txt
streamlit
pillow
opencv-python
numpy
rembg
ultralytics